+++ src/main/resources/com/example/marvelsnaptracker/deck-list-view.fxml
@@ -6,7 +6,7 @@
 <?import java.net.URL?>
 <ScrollPane xmlns:fx="http://javafx.com/fxml/1"
             xmlns="http://javafx.com/javafx/17.0.2-ea"
-            fx:controller="com.example.marvelsnaptracker.contoller.DeckListViewController"
+            fx:controller="com.example.marvelsnaptracker.controller.DeckListViewController"
             stylesheets="@styles/deck-list-view.css"
             styleClass="deck-view-container" fitToWidth="true">
