+++ src/main/resources/com/example/marvelsnaptracker/deck-view.fxml
@@ -7,7 +7,7 @@
 
 <VBox xmlns="http://javafx.com/javafx/17.0.2-ea"
       xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.example.marvelsnaptracker.contoller.DeckViewController"
+      fx:controller="com.example.marvelsnaptracker.controller.DeckViewController"
       stylesheets="@styles/deck-view.css"
       styleClass="deck-container" >
