+++ src/main/resources/com/example/marvelsnaptracker/main-view.fxml
@@ -3,11 +3,10 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<?import java.net.URL?>
 <?import javafx.geometry.Insets?>
 <VBox xmlns="http://javafx.com/javafx"
       xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.marvelsnaptracker.contoller.MainViewController"
+      fx:controller="com.example.marvelsnaptracker.controller.MainViewController"
       stylesheets="@styles/main-view.css"
       styleClass="main-container">
